Initial commit
I have one btrfs filesystem that I snapshot regularly, so I have a
sequence of snapshots (named after date) where one is an incremental
change on top of the previous one.

Now I created a second filesystem for backup, to guard against btrfs
bugs, so it should be an isolated filesystem. I can rsync the current
state, but I also want to replicate the snapshots. Getting that synced
requires some care to preserve sharing between snapshots.

There is btrfs send/receive, but I do not want to use those. They are
too low-level, I fear that a corruption on one end would cascade and
render both filesystems unusable. I need the two file systems to be as
isolated as possible. So basically, what I need to do is compute the
diff between two snapshots on the sending side, and then apply the diff
on the receiving side in such a way to take advantage of reflinking.

My first attempt was to do this with rsync, with some combination of
flags that works reasonably, but still it will re-write moved files for
example. There is no reflink support in rsync (there is a patch floating
around, but no activity on the bug since 2015), and my hypothesis is
that reliably detecting renames and reflinking them is more important
than incremental updates to existing files. That is, most of my data is
immutable, write-once. And I tend to move those around more than I
modify existing files. So my hypothesis is that I can do better than
rsync, by detecting renames and taking advantage of reflinks, even if I
don't have rsync's smart deltas at the file level.

Step one for this would be to enumerate the files in the two snapshots,
and compute a diff.