Run reflinker, debug rsync flags
So half of my files go missing silently now, and I have discovered that
with --compare-dest, rsync will delete some of the files that have not
changed between snapshots. It does not print that it deletes them with
--info=del. Running a second time with the same flags finds no
differences. But running a second time without --compare-dest, it does
pick up the missing files.