Remove dangerous rsync option
I thought that --compare-dest merely provides more context for delta
updates, but that is not the only thing it does: it also causes target
files that are already present in the --compare-dest dir to be deleted
at the target side (since rsync 3.1). That is dangerous here, and that
was why my files were disappearing.