Prefer syncing later snapshots first
Previously I had this idea of using the "closest" snapshot first, with
building forward in time counting twice as far as building backwards in
time. But now I am in this situation:

  Source has   Destination has
  ----------   ---------------
  d
  d-1
  d-2
  d-3
  ...
  d-n+1
  d-n          d-n
  d-n-1        d-n-1
  d-n-2        d-n-2
  ...          ...

And that would mean it would start to sync d-n+1, using d-n as a base,
and then build the entire thing forward. But I'd rather build d, using
d-n as a base, and then build backward from there. That would lead to
less fragmented files in the "d" snapshot for files that do change, and
for write-once files it does not matter. For mostly append-only data, it
means the initial sync will transfer a lot of data, and the following
syncs are mostly just deleting things that weren't yet present in that
snapshot to make a new snapshot.